* LPController.m           int -> NSInteger transition
git-svn-id: svn://svn.savannah.nongnu.org/gap/trunk@2226 d25b4930-3292-e311-b074-0010c6de5da8

+++ user-apps/Games/LapisPuzzle/LPController.m
@@ -229,15 +229,15 @@
 												 repeats:YES]);
 }
 
-- (int) browser:(NSBrowser *)sender numberOfRowsInColumn:(int)column
+- (NSInteger) browser:(NSBrowser *)sender numberOfRowsInColumn:(NSInteger)column
 {
 	return [attackArray count];
 }
 
 - (void) browser: (NSBrowser *)sender
  willDisplayCell: (NSBrowserCell *)cell
-		   atRow: (int)row
-		  column: (int)column
+		   atRow: (NSInteger)row
+		  column: (NSInteger)column
 {
 	[cell setTitle:[attackArray objectAtIndex:row]];
 	[cell setLeaf:YES];
